restructure, separate landing page
